Initial commit of mysql-connector-cpp. Fixes T4281
Summary:
NOTE:
Kudos for this patch is due to @herbmillerjr. Just
trying to finish T243, as there has been no response from
him since november 2017, and expecting that many users can
be interested in having mysql-workbench in the repos.
I've simple changed the formating of the package.yml,
as asked by @JoshStrobl in D1401 and completed a test
to ensure that it works.

Test Plan:
- Compiled and run this test file:
https://dev.mysql.com/doc/connector-cpp/en/connector-cpp-examples-complete-example-1.html

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake, JoshStrobl, herbmillerjr

Differential Revision: https://dev.solus-project.com/D3052